+++ src/rdbus/Manager.hpp
@@ -26,7 +26,7 @@ private:
     std::string name;
     Tasks tasks;
     Output output;
-    std::optional< Data > run( std::unique_ptr< tasks::Task >& task );
+    std::list< Data > run( std::unique_ptr< tasks::Task >& task );
     void send( const std::list< rdbus::Data >& list );
 };